Cropping WIP.
git-svn-id: http://ebookdroid.googlecode.com/svn/trunk@577 714301b5-5bf2-3aa6-4ccc-ecebab0412e3

+++ src/org/ebookdroid/core/Page.java
@@ -89,7 +89,7 @@ public class Page {
         return false;
     }
 
-    public boolean setAspectRatio(final int width, final int height) {
+    public boolean setAspectRatio(final float width, final float  height) {
         return setAspectRatio((width / pageType.getWidthScale()) / height);
     }
